+++ tf-modules/cumulus/variables.tf
@@ -487,3 +487,9 @@ variable "non_complete_execution_payload_timeout" {
   type        = number
   default     = 30
 }
+
+variable "archive_api_url" {
+  type        = string
+  default     = null
+  description = "If not specified, the value of the Backend (Archive) API Gateway endpoint is used"
+}